Day9

+++ kotlin/src/main/kotlin/Day5.kt
@@ -4,11 +4,11 @@ import kotlin.collections.ArrayList
 class Day5 {
     companion object {
 
-        fun solveDay5Part1(input: List<Int>) {
+        fun solveDay5Part1(input: List<Long>) {
             IntCodeComputer(input).run(ArrayList(), null)
         }
 
-        fun solveDay5Part2(input: List<Int>) {
+        fun solveDay5Part2(input: List<Long>) {
             solveDay5Part1(input)
         }
     }
@@ -18,9 +18,9 @@ fun main() {
 
     val input = Scanner(Day5::class.java.getResourceAsStream("Day5Input")).useDelimiter("[,\n]")
 
-    val inputList: MutableList<Int> = ArrayList()
-    while (input.hasNextInt()) {
-        inputList.add(input.nextInt())
+    val inputList: MutableList<Long> = ArrayList()
+    while (input.hasNextLong()) {
+        inputList.add(input.nextLong())
     }
 
     Day5.solveDay5Part1(inputList)
